Rename console script modules to cli
